requirements.txt
customtkinter==5.2.2
icrawler==0.6.9
Pillow==10.4.0
